+++ src/buffetBotJsApi/package.json
@@ -13,7 +13,7 @@
 	"type": "module",
 	"scripts": {
 		"serve": "pnpm compile && pnpm execute",
-		"compile": "pnpm exec tsc",
+		"compile": "pnpm tsc --build",
 		"execute": "node ./dist/buffetBotJsApi/main.js"
 	},
 	"dependencies": {
