+++ src/danielTurcichComApi/package.json
@@ -13,7 +13,7 @@
 	"type": "module",
 	"scripts": {
 		"serve": "pnpm compile && pnpm execute",
-		"compile": "pnpm exec tsc",
+		"compile": "pnpm tsc --build",
 		"execute": "node ./dist/danielTurcichComApi/main.js"
 	},
 	"dependencies": {
