+++ tsconfig.json
@@ -24,10 +24,12 @@
 		"allowSyntheticDefaultImports": true,
 
 		"baseUrl": ".",
+		"rootDir": ".",
 		"paths": {
 			"@shared/*": ["src/shared/*"]
 		},
 		"lib": ["esnext", "scripthost"]
 	},
+	"include": ["**/*.ts", "./src/types/global.d.ts"],
 	"exclude": ["node_modules"]
 }
